AG-12895 Join SeriesAreaManager and SeriesAreaFocusManager
